pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urban-air-quality-intelligence-system"
version = "0.1.0"
description = "Urban Air Quality Intelligence System: collect, analyze, and predict urban air pollution using real data."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ankit Reddy", email = "ankitreddy@example.com" }]

dependencies = [
  "python-dotenv>=1.0.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "requests>=2.32",
  "scikit-learn>=1.4",
  "prophet>=1.1.5",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "plotly>=5.17",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "folium>=0.15",
  "streamlit>=1.34",
  "streamlit-folium>=0.14",
  "joblib>=1.3",
  "tqdm>=4.66",
  "beautifulsoup4>=4.12"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "flake8>=6.1",
  "black>=24.0",
  "isort>=5.13",
  "mypy>=1.9",
  "types-requests>=2.32.0.20240712",
  "types-python-dateutil>=2.9.0.20240316"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
py_version = 311
